refactor(persistence): move conversation group/display-order migration guards into persistence/
