+++ assistant/src/persistence/conversation-group-migration.ts
@@ -8,9 +8,9 @@
  * than the formal migrations/ pipeline.
  */
 
-import { rawExec, rawGet, rawRun } from "../persistence/raw-query.js";
 import { getLogger } from "../util/logger.js";
 import { ensureDisplayOrderMigration } from "./conversation-display-order-migration.js";
+import { rawExec, rawGet, rawRun } from "./raw-query.js";
 const log = getLogger("conversation-store");
 
 function isDuplicateColumnError(err: unknown): boolean {
