+++ assistant/src/persistence/group-crud.ts
@@ -7,8 +7,8 @@
 
 import { v4 as uuid } from "uuid";
 
-import { ensureGroupMigration } from "../memory/conversation-group-migration.js";
 import { rawAll, rawExec, rawGet, rawRun } from "../persistence/raw-query.js";
+import { ensureGroupMigration } from "./conversation-group-migration.js";
 export interface ConversationGroupRow {
   id: string;
   name: string;
